Aula 08 Tratamento de erros e use History

+++ src/pages/Repositories/index.js
@@ -1,13 +1,23 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useHistory  } from 'react-router';
+
+
 
 export default function Repositories() {
+    const history = useHistory();
     const [ repositories, setRepositories ] = useState([]); 
+    
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if(repositoriesName != null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            history.push('/')
+        }
+        
 
     }, []);
 
@@ -21,6 +31,7 @@ export default function Repositories() {
                     )
                 })}
             </S.List>
+            <S.LinkHome to='/'>Voltar</S.LinkHome>
         </S.Container>
     )
 }
